feat: rename util to internal
