+++ internal/dao/chainAccount.go
@@ -8,8 +8,8 @@ import (
 	"github.com/itering/subscan/internal/model"
 	"github.com/itering/subscan/internal/substrate"
 	"github.com/itering/subscan/internal/substrate/rpc"
-	"github.com/itering/subscan/util"
-	"github.com/itering/subscan/util/ss58"
+	"github.com/itering/subscan/internal/util"
+	"github.com/itering/subscan/internal/util/ss58"
 	"github.com/jinzhu/gorm"
 	"github.com/shopspring/decimal"
 )
