+++ internal/dao/chainLog.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/itering/subscan/internal/model"
 	"github.com/itering/subscan/internal/substrate/storage"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"strings"
 )
