+++ internal/dao/chainTransaction.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 	"github.com/itering/subscan/internal/model"
 	"github.com/itering/subscan/internal/substrate"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"github.com/shopspring/decimal"
 	"strings"
 )
