+++ internal/dao/extrinsicError.go
@@ -3,7 +3,7 @@ package dao
 import (
 	"github.com/itering/subscan/internal/model"
 	"github.com/itering/subscan/internal/substrate"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"strings"
 )
