+++ internal/server/http/common.go
@@ -4,7 +4,7 @@ import (
 	"github.com/go-kratos/kratos/pkg/log"
 	bm "github.com/go-kratos/kratos/pkg/net/http/blademaster"
 	"github.com/itering/subscan/internal/substrate"
-	"github.com/itering/subscan/util/ss58"
+	"github.com/itering/subscan/internal/util/ss58"
 	"net/http"
 	"time"
 )
