+++ internal/service/balances/balances.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"github.com/itering/subscan/internal/dao"
 	"github.com/itering/subscan/internal/model"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"github.com/shopspring/decimal"
 )
