+++ internal/service/event.go
@@ -7,7 +7,7 @@ import (
 	"github.com/itering/subscan/internal/model"
 	"github.com/itering/subscan/internal/service/balances"
 	"github.com/itering/subscan/internal/service/system"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"github.com/shopspring/decimal"
 	"strings"
 )
