+++ internal/service/scan/scan.go
@@ -5,8 +5,8 @@ import (
 	"github.com/itering/subscan/internal/dao"
 	"github.com/itering/subscan/internal/model"
 	"github.com/itering/subscan/internal/substrate"
-	"github.com/itering/subscan/util"
-	"github.com/itering/subscan/util/ss58"
+	"github.com/itering/subscan/internal/util"
+	"github.com/itering/subscan/internal/util/ss58"
 	"github.com/json-iterator/go"
 	"regexp"
 )
