+++ internal/service/service.go
@@ -7,7 +7,7 @@ import (
 	"github.com/itering/subscan/internal/dao"
 	"github.com/itering/subscan/internal/service/scan"
 	"github.com/itering/subscan/internal/substrate/metadata"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"io/ioutil"
 	"strings"
 )
