+++ internal/substrate/extrinsic.go
@@ -5,7 +5,7 @@ import (
 	scalecodec "github.com/freehere107/go-scale-codec"
 	"github.com/freehere107/go-scale-codec/types"
 	"github.com/itering/subscan/internal/substrate/metadata"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 )
 
 // Extrinsic decode
