+++ internal/substrate/log.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/freehere107/go-scale-codec/types"
 	"github.com/itering/subscan/internal/substrate/storage"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 )
 
 const (
