+++ internal/substrate/metadata/metadata.go
@@ -3,7 +3,7 @@ package metadata
 import (
 	"github.com/freehere107/go-scale-codec"
 	"github.com/freehere107/go-scale-codec/types"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"strings"
 )
