+++ internal/substrate/storage/decode.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 
 	"github.com/freehere107/go-scale-codec/types"
-	"github.com/itering/subscan/util"
+	"github.com/itering/subscan/internal/util"
 	"github.com/shopspring/decimal"
 )
