+++ internal/util/ecode.go
