+++ internal/util/hex.go
