+++ internal/util/hex_test.go
