+++ internal/util/interface.go
