+++ internal/util/result/fallback.go
