+++ internal/util/result/result.go
