+++ internal/util/ss58/ss58.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"encoding/binary"
 	"github.com/freehere107/go-scale-codec/types"
-	"github.com/itering/subscan/util"
-	"github.com/itering/subscan/util/base58"
+	"github.com/itering/subscan/internal/util"
+	"github.com/itering/subscan/internal/util/base58"
 	"golang.org/x/crypto/blake2b"
 )
