+++ internal/util/string.go
